pyproject.toml
[project]
name = "geo-testing"
version = "0.1.0"
description = ""
authors = [
    {name = "Benjamin Marler",email = "b.marler@northeastern.edu"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright (>=1.55.0,<2.0.0)",
    "pymongo (>=4.15.3,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "geo_testing", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
